requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
fastapi==0.110.0
h11==0.14.0
idna==3.6
iso8601==1.1.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
sniffio==1.3.1
starlette==0.36.3
tortoise-orm==0.20.0
typing_extensions==4.10.0
uvicorn==0.29.0
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
fastapi==0.110.0
h11==0.14.0
idna==3.6
iso8601==1.1.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
sniffio==1.3.1
starlette==0.36.3
tortoise-orm==0.20.0
typing_extensions==4.10.0
uvicorn==0.29.0
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
dictdiffer==0.9.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
iso8601==1.1.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
sniffio==1.3.1
starlette==0.36.3
tomlkit==0.12.4
tortoise-orm==0.20.0
typing_extensions==4.10.0
uvicorn==0.29.0
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
dictdiffer==0.9.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
iso8601==1.1.0
packaging==24.0
pluggy==1.4.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pypika-tortoise==0.1.6
pytest==8.1.1
python-dotenv==1.0.1
pytz==2024.1
sniffio==1.3.1
starlette==0.36.3
tomlkit==0.12.4
tortoise-orm==0.20.0
typing_extensions==4.10.0
uvicorn==0.29.0
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
click==8.1.7
dictdiffer==0.9.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
iso8601==1.1.0
packaging==24.0
pluggy==1.4.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pypika-tortoise==0.1.6
pytest==8.1.1
python-dotenv==1.0.1
pytz==2024.1
sniffio==1.3.1
starlette==0.36.3
tomlkit==0.12.4
tortoise-orm==0.20.0
typing_extensions==4.10.0
uvicorn==0.29.0
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dictdiffer==0.9.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
iso8601==1.1.0
packaging==24.0
pluggy==1.4.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pypika-tortoise==0.1.6
pytest==8.1.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
tomlkit==0.12.4
tortoise-orm==0.20.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dictdiffer==0.9.0
fastapi==0.110.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
iso8601==1.1.0
packaging==24.0
pluggy==1.4.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pypika-tortoise==0.1.6
pytest==8.1.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
tomlkit==0.12.4
tortoise-orm==0.20.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
